refactor: move provider types to @libztbs/types for zero-dependency foundation
InferredProvider, ExtendedProvider型をai-detectorからtypesに移動。
@libztbs/typesがゼロ依存となり、依存グラフの最下層に正しく位置する。
